Rename outputs -> results
