+++ synth/results/visits_age_range_count.sql
