+++ synth/results/visits_count.sql
