+++ synth/results/visits_gender_count.sql
